+++ semgrep/semgrep/config_resolver.py
@@ -1,6 +1,4 @@
-import base64
 import os
-import shutil
 import sys
 import tarfile
 import tempfile
@@ -18,6 +16,7 @@ from semgrep.constants import ID_KEY
 from semgrep.constants import RULES_KEY
 from semgrep.constants import SEMGREP_USER_AGENT
 from semgrep.constants import YML_EXTENSIONS
+from semgrep.rule_lang import parse_yaml_with_spans
 from semgrep.util import debug_print
 from semgrep.util import is_url
 from semgrep.util import print_error
@@ -90,19 +89,20 @@ def parse_config_at_path(
         config_id = str(loc).replace(str(base_path), "")
     try:
         with loc.open() as f:
-            return parse_config_string(config_id, f.read())
+            return parse_config_string(config_id, f.read(), str(loc))
     except FileNotFoundError:
         print_error(f"YAML file at {loc} not found")
         return {str(loc): None}
 
 
 def parse_config_string(
-    config_id: str, contents: str
+    config_id: str, contents: str, filename: Optional[str] = None,
 ) -> Dict[str, Optional[Dict[str, Any]]]:
-    import yaml  # here for faster startup times
+    import yaml.parser
+    import yaml.scanner
 
     try:
-        return {config_id: yaml.safe_load(contents)}
+        return {config_id: parse_yaml_with_spans(contents=contents, filename=filename)}
     except yaml.parser.ParserError as se:
         print_error(f"Invalid yaml file {config_id}:\n{indent(str(se))}")
         return {config_id: None}
@@ -186,7 +186,9 @@ def download_config(config_url: str) -> Dict[str, Optional[Dict[str, Any]]]:
             content_type = r.headers.get("Content-Type")
             if content_type and "text/plain" in content_type:
                 print_msg(SCANNING_MESSAGE)
-                return parse_config_string("remote-url", r.content.decode("utf-8"))
+                return parse_config_string(
+                    "remote-url", r.content.decode("utf-8"), config_url
+                )
             elif content_type and content_type == "application/x-gzip":
                 with tempfile.TemporaryDirectory() as fname:
                     with tarfile.open(fileobj=r.raw, mode="r:gz") as tar: