+++ semgrep/semgrep/dump_ast.py
@@ -1,11 +1,13 @@
 import subprocess
 import tempfile
+from pathlib import Path
 from typing import List
 from typing import Optional
 
 import semgrep.config_resolver
 from semgrep.constants import PLEASE_FILE_ISSUE_TEXT
 from semgrep.constants import SEMGREP_PATH
+from semgrep.error import SemgrepException
 from semgrep.util import print_error
 from semgrep.util import print_error_exit
 
@@ -14,16 +16,27 @@ def dump_parsed_ast(
     to_json: bool, language: str, pattern: Optional[str], targets_str: List[str]
 ) -> None:
     targets = semgrep.config_resolver.resolve_targets(targets_str)
+    if pattern is None and len(targets) != 1:
+        print_error_exit(
+            f"exactly one target file is required with this option (got {targets})"
+        )
+    print(parsed_ast(to_json, language, pattern, targets))
+
+
+def parsed_ast(
+    to_json: bool, language: str, pattern: Optional[str], targets: List[Path]
+) -> str:
 
     with tempfile.NamedTemporaryFile("w") as fout:
-        args = []
         if pattern:
             fout.write(pattern)
             fout.flush()
             args = ["-lang", language, "-dump_pattern", fout.name]
         else:
             if len(targets) != 1:
-                print_error_exit("exactly one target file is required with this option")
+                raise ValueError(
+                    f"When pattern is not set, exactly one target file is required. (Got {targets})"
+                )
             target = targets[0]
             args = ["-lang", language, "-dump_ast", str(target)]
 
@@ -36,4 +49,4 @@ def dump_parsed_ast(
         except subprocess.CalledProcessError as ex:
             print_error(f"error invoking semgrep with:\n\t{' '.join(cmd)}\n{ex}")
             print_error_exit(f"\n\n{PLEASE_FILE_ISSUE_TEXT}")
-        print(output.decode())
+        return output.decode()